pyproject.toml
[project]
name = "mcf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Wong Ding Feng", email = "dingfengwong@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.3",
    "httpx>=0.28.1",
    "numpy>=2.2.0",
    "pandas>=2.3.3",
    "polars>=1.36.1",
    "pyarrow>=22.0.0",
    "pypdf>=6.0.0",
    "pydantic>=2.12.5",
    "rich>=14.2.0",
    "sentence-transformers>=5.0.0",
    "tenacity>=9.1.2",
    "typer>=0.21.0",
    "python-docx>=1.2.0",
]

[project.scripts]
mcf = "mcf.cli.cli:main"

[build-system]
requires = ["uv_build>=0.8.23,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "black>=25.12.0",
    "build>=1.3.0",
    "mypy>=1.19.1",
    "pip>=25.3",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
    "sphinx>=9.0.4",
    "twine>=6.2.0",
]
